chore(mcp): move statuspage to new architecture
