+++ front/lib/actions/mcp_internal_actions/servers/index.ts
@@ -54,7 +54,7 @@ import { default as slackServer } from "@app/lib/actions/mcp_internal_actions/se
 import { default as slideshowServer } from "@app/lib/actions/mcp_internal_actions/servers/slideshow";
 import { default as soundStudio } from "@app/lib/actions/mcp_internal_actions/servers/sound_studio";
 import { default as speechGenerator } from "@app/lib/actions/mcp_internal_actions/servers/speech_generator";
-import { default as statuspageServer } from "@app/lib/actions/mcp_internal_actions/servers/statuspage";
+import { default as statuspageServer } from "@app/lib/api/actions/servers/statuspage";
 import { default as tablesQueryServerV2 } from "@app/lib/actions/mcp_internal_actions/servers/tables_query";
 import { default as toolsetsServer } from "@app/lib/actions/mcp_internal_actions/servers/toolsets";
 import { default as ukgReadyServer } from "@app/lib/actions/mcp_internal_actions/servers/ukg_ready";
