+++ front/lib/api/actions/servers/statuspage/rendering.ts
@@ -2,7 +2,7 @@ import type {
   StatuspageComponent,
   StatuspageIncident,
   StatuspagePage,
-} from "@app/lib/actions/mcp_internal_actions/servers/statuspage/types";
+} from "@app/lib/api/actions/servers/statuspage/types";
 
 function formatStatus(status: string): string {
   return status.replace(/_/g, " ").replace(/\b\w/g, (c) => c.toUpperCase());
